Moving things around, cleaning up silliness, and making it work.
