+++ classes/lister/UserFormatList.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * DynamicPageList3
+ * DPL UserFormatList Class
+ *
+ * @license		GPL-2.0-or-later
+ * @package		DynamicPageList3
+ *
+ **/
+
+namespace DPL\Lister;
+
+class UserFormatList extends Lister {
+	/**
+	 * Listing style for this class.
+	 *
+	 * @var		constant
+	 */
+	public $style = parent::LIST_USERFORMAT;
+}
+?>
