+++ extension.json
@@ -38,13 +38,13 @@
 		"DPL\\DynamicPageList": "classes/DynamicPageList.php",
 		"DPL\\Logger": "classes/Logger.php",
 		"DPL\\ListMode": "classes/ListMode.php",
-		"DPL\\List\\List": "classes/list/List.php",
-		"DPL\\List\\DefinitionList": "classes/list/DefinitionList.php",
-		"DPL\\List\\GalleryList": "classes/list/GalleryList.php",
-		"DPL\\List\\HeadingList": "classes/list/HeadingList.php",
-		"DPL\\List\\InlineList": "classes/list/InlineList.php",
-		"DPL\\List\\OrderedList": "classes/list/OrderedList.php",
-		"DPL\\List\\UnorderedList": "classes/list/UnorderedList.php",
+		"DPL\\Lister\\Lister": "classes/lister/Lister.php",
+		"DPL\\Lister\\DefinitionList": "classes/lister/DefinitionList.php",
+		"DPL\\Lister\\GalleryList": "classes/lister/GalleryList.php",
+		"DPL\\Lister\\HeadingList": "classes/lister/HeadingList.php",
+		"DPL\\Lister\\InlineList": "classes/lister/InlineList.php",
+		"DPL\\Lister\\OrderedList": "classes/lister/OrderedList.php",
+		"DPL\\Lister\\UnorderedList": "classes/lister/UnorderedList.php",
 		"DPL\\LST": "classes/LST.php",
 		"DPL\\Parameters": "classes/Parameters.php",
 		"DPL\\ParametersData": "classes/ParametersData.php",
